+++ test/complextensor/test_complextensor.py
@@ -29,7 +29,7 @@ except ImportError:
         Variant,
     )
 
-from torch.complextensor.ops.common import ComplexTensorMode
+from torch.complextensor._ops.common import ComplexTensorMode
 from torch.testing._internal.common_device_type import (
     instantiate_device_type_tests,
     OpDTypes,
