+++ torch/complextensor/__init__.py
@@ -1,5 +1,5 @@
 from .core import ComplexTensor
-from .ops import ComplexTensorMode, is_complex_tensor
+from ._ops import ComplexTensorMode, is_complex_tensor
 
 
 __all__ = ["ComplexTensor", "ComplexTensorMode", "is_complex_tensor"]
